fix: update preset usage on react-router implementation
fixed documentation
