+++ examples/react-router/pwa-simple-sw/react-router.config.ts
@@ -2,5 +2,5 @@ import { ReactRouterPreset } from '@vite-pwa/remix'
 
 export default {
   ssr: process.env.SPA !== 'true',
-  presets: [ReactRouterPreset({} as any)()],
+  presets: [ReactRouterPreset],
 }
